refactor graph_view
